[QOLDEV-915] fix integration with QA plugin
- Previously we detected the QA plugin presence by checking for it in the config, but this has stopped working,
possibly as part of the CKAN 2.10 upgrade. Change it to instead attempt the necessary imports and rely on that.

NB This will cause ckanext-data-qld to implement the IQA interface if QA is installed but not enabled.
However, that should not be problematic, since the interface will simply do nothing if QA is not running.